Reversal of Elements
It reverses the order of the values in data, then prints it out.
In the first version of the program there is only one array and its values are reversed.
In the second version of the program  the first array data is not changed. The second array result gets the elements of data in reversed order.